# api/openapi.yaml
openapi: 3.1.0
info:
  title: go-service-main API
  version: 0.4.0
  description: Single-tenant administration API compatible with vue3-element-admin.
security:
  - bearerAuth: []
paths:
...
  /api/v1/depts:
    get:
      operationId: listDepartments
      parameters:
        - { in: query, name: keywords, schema: { type: string } }
        - { in: query, name: status, schema: { type: integer, enum: [0, 1] } }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    post:
      operationId: createDepartment
      requestBody: { $ref: '#/components/requestBodies/DepartmentForm' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
...
  /api/v1/depts/{deptId}/form:
    parameters:
      - $ref: '#/components/parameters/DepartmentID'
    get:
      operationId: getDepartmentForm
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/depts/{deptId}:
    parameters:
      - $ref: '#/components/parameters/DepartmentID'
    put:
      operationId: updateDepartment
      requestBody: { $ref: '#/components/requestBodies/DepartmentForm' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    delete:
      operationId: deleteDepartments
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
...
  /api/v1/notices/{noticeId}/form:
    parameters:
      - { in: path, name: noticeId, required: true, schema: { type: string } }
    get:
      operationId: getNoticeForm
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/notices/{noticeId}/detail:
    parameters:
      - { in: path, name: noticeId, required: true, schema: { type: string } }
    get:
      operationId: getNoticeDetail
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/notices/{noticeId}:
    parameters:
      - { in: path, name: noticeId, required: true, schema: { type: string } }
    put:
      operationId: updateNotice
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    delete:
      operationId: deleteNotices
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/notices/{noticeId}/publish:
    parameters:
      - { in: path, name: noticeId, required: true, schema: { type: string } }
    put:
      operationId: publishNotice
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/notices/{noticeId}/revoke:
    parameters:
      - { in: path, name: noticeId, required: true, schema: { type: string } }
    put:
      operationId: revokeNotice
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    DepartmentID:
      in: path
      name: deptId
      required: true
      schema: { type: string }
...
  requestBodies:
...
    DepartmentForm:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DepartmentForm' }
...
  responses:
    AdminAPISuccess:
      description: Successful administration API response.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminAPIResult' }
...
  schemas:
...
    DepartmentForm:
      type: object
      required: [name, code, status]
      properties:
        name: { type: string }
        code: { type: string }
        parentId: { type: string }
        sort: { type: integer }
        status: { type: integer, enum: [0, 1] }
    AdminAPIResult:
      type: object
      required: [code, msg, data]
      properties:
        code: { type: string }
        msg: { type: string }
        data: {}
        requestId: { type: string }